Refactor spline functions to a different file
