+++ lib/mayaUsd/fileio/utils/writeUtil.cpp
@@ -24,13 +24,6 @@
 #include <pxr/base/gf/rotation.h>
 #include <pxr/base/tf/envSetting.h>
 #include <pxr/base/tf/token.h>
-#if PXR_VERSION >= 2411
-#include <pxr/base/ts/knot.h>
-#include <pxr/base/ts/spline.h>
-#include <pxr/base/ts/tangentConversions.h>
-#include <pxr/base/ts/typeHelpers.h>
-#include <pxr/base/ts/types.h>
-#endif
 #include <pxr/base/vt/types.h>
 #include <pxr/base/vt/value.h>
 #include <pxr/pxr.h>
@@ -1279,149 +1272,5 @@ std::vector<double> UsdMayaWriteUtil::GetTimeSamples(
     std::sort(samples.begin(), samples.end());
     return samples;
 }
-
-#if PXR_VERSION >= 2411
-static TsExtrapMode _ConvertExtrapolationType(MFnAnimCurve::InfinityType mayaExtrapolation)
-{
-    switch (mayaExtrapolation) {
-    case MFnAnimCurve::InfinityType::kLinear: return TsExtrapMode::TsExtrapLinear;
-    case MFnAnimCurve::InfinityType::kCycle: return TsExtrapMode::TsExtrapLoopReset;
-    case MFnAnimCurve::InfinityType::kOscillate: return TsExtrapMode::TsExtrapLoopOscillate;
-    case MFnAnimCurve::InfinityType::kCycleRelative: return TsExtrapMode::TsExtrapLoopRepeat;
-    case MFnAnimCurve::InfinityType::kConstant:
-    default: return TsExtrapMode::TsExtrapHeld;
-    }
-}
-
-static TsInterpMode _ConvertMayaTanTypeToUsdTanType(MFnAnimCurve::TangentType mayaTangentType)
-{
-    switch (mayaTangentType) {
-    case MFnAnimCurve::TangentType::kTangentStep:
-    case MFnAnimCurve::TangentType::kTangentStepNext: return TsInterpHeld;
-    case MFnAnimCurve::TangentType::kTangentLinear: return TsInterpLinear;
-    default: return TsInterpCurve;
-    }
-}
-
-TsKnotMap UsdMayaWriteUtil::GetKnotsFromMayaCurve(
-    const MFnDependencyNode& depNode,
-    const MString&           name,
-    float                    scaling)
-{
-    TsKnotMap knots;
-
-    MStatus status;
-    depNode.attribute(name, &status);
-    CHECK_MSTATUS_AND_RETURN(status, knots)
-    MPlug plug = depNode.findPlug(name, true, &status);
-    CHECK_MSTATUS_AND_RETURN(status, knots)
-
-    // get the animation curve for the given maya attribute if there's one
-    MFnAnimCurve flAnimCurve(plug, &status);
-    if (status != MStatus::kSuccess) {
-        return knots;
-    }
-
-    auto numKeys = flAnimCurve.numKeys();
-    for (unsigned int k = 0; k < numKeys; ++k) {
-        auto time = flAnimCurve.time(k).value();
-
-        auto   value = flAnimCurve.value(k);
-        MTime  convert(1.0, MTime::kSeconds);
-        double inTangentX, inTangentY;
-        double outTangentX, outTangentY;
-        flAnimCurve.getTangent(k, inTangentX, inTangentY, true);
-        flAnimCurve.getTangent(k, outTangentX, outTangentY, false);
-
-        // This was taken from the .getTangent() docs:
-        // Need to multiply the value with the time unit conversion factor
-        inTangentX *= convert.as(MTime::uiUnit());
-        outTangentX *= convert.as(MTime::uiUnit());
-
-        TsTime inTime {}, outTime {};
-        float  inSlope = 0.f, outSlope = 0.f;
-
-        // Converting from maya tangent to standard (Usd) tangent:
-        // Usd tangents are specified by slope and length and Slopes are "rise over run": height
-        // divided by length.
-        // Maya tangents are specified by height and length. Height and length
-        // are both specified multiplied by 3 Heights are positive for upward-sloping post-tangents,
-        // and negative for upward-sloping pre-tangents.
-        TsConvertToStandardTangent(
-            static_cast<float>(inTangentX),
-            static_cast<float>(inTangentY),
-            true,
-            true,
-            true,
-            &inTime,
-            &inSlope);
-
-        TsKnot     knot(Ts_GetType<float>());
-        const auto outTanType = flAnimCurve.outTangentType(k);
-        const auto convertedValue = static_cast<float>(value) * scaling;
-
-        // Deal with the case where slope would be infinite,
-        // Because when there's a step on the curve is discontinuous
-        if (outTanType == MFnAnimCurve::kTangentStepNext) {
-            // Maya's step next is a special case where the value jumps to the next key's value.
-            // If this is the last key, then set the value to the current value, making it behave
-            // like a step.
-            knot.SetPreValue(convertedValue);
-            if (k + 1 < numKeys) {
-                knot.SetValue(static_cast<float>(flAnimCurve.value(k + 1)) * scaling);
-            } else {
-                knot.SetValue(convertedValue);
-            }
-        } else if (outTanType == MFnAnimCurve::kTangentStep) {
-            // no need to convert tangent in this case because it is 0 for the step.
-            knot.SetValue(static_cast<float>(value) * scaling);
-        } else {
-            TsConvertToStandardTangent(
-                static_cast<float>(outTangentX),
-                static_cast<float>(outTangentY),
-                true,
-                true,
-                false,
-                &outTime,
-                &outSlope);
-            knot.SetValue(static_cast<float>(value) * scaling);
-        }
-
-        knot.SetTime(time);
-        knot.SetPostTanSlope(outSlope);
-        knot.SetPreTanSlope(inSlope);
-        knot.SetPostTanWidth(outTime);
-        knot.SetPreTanWidth(inTime);
-        knot.SetNextInterpolation(_ConvertMayaTanTypeToUsdTanType(outTanType));
-
-        knots.insert(knot);
-    }
-
-    return knots;
-}
-
-TsSpline
-UsdMayaWriteUtil::GetSplineFromMayaCurve(const MFnDependencyNode& depNode, const MString& name)
-{
-    auto spline = TsSpline(Ts_GetType<float>());
-
-    MStatus status;
-    depNode.attribute(name, &status);
-    CHECK_MSTATUS_AND_RETURN(status, spline)
-    MPlug plug = depNode.findPlug(name, true, &status);
-    CHECK_MSTATUS_AND_RETURN(status, spline)
-
-    // get the animation curve for the given maya attribute
-    MFnAnimCurve flAnimCurve(plug, &status);
-    CHECK_MSTATUS_AND_RETURN(status, spline)
-
-    TsExtrapolation preExtrapolation(_ConvertExtrapolationType(flAnimCurve.preInfinityType()));
-    TsExtrapolation postExtrapolation(_ConvertExtrapolationType(flAnimCurve.postInfinityType()));
-    spline.SetPreExtrapolation(preExtrapolation);
-    spline.SetPostExtrapolation(postExtrapolation);
-
-    return spline;
-}
-#endif
 
 PXR_NAMESPACE_CLOSE_SCOPE
